Update the open source portal for Fsq.io.
This creates a new custom category for Fsq.io, Foursquare's
new way of deploying OSS. It is a repo that is directly
synced with our monorepo, and is the new home for all
of the projects that were previously listed under Server.

This also deletes an old Blackberry project from this list.